requirements-dev.txt
appdirs==1.4.4
APScheduler==3.6.3
astroid==2.4.2
black==20.8b1
certifi==2020.12.5
cffi==1.14.4
click==7.1.2
cryptography==3.3.1
decorator==4.4.2
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
pycparser==2.20
pylint==2.6.0
python-dateutil==2.8.1
python-telegram-bot==13.1
pytz==2020.5
regex==2020.11.13
six==1.15.0
toml==0.10.2
tornado==6.1
typed-ast==1.4.2
typing-extensions==3.7.4.3
tzlocal==2.1
wrapt==1.12.1
